Javadoc and rename job
