+++ src/main/groovy/org/mockito/release/internal/util/ExposedForTesting.java
@@ -1,0 +1,7 @@
+package org.mockito.release.internal.util;
+
+/**
+ * Informs that the method or field is exposed for testing.
+ */
+public @interface ExposedForTesting {
+}
